fix: dfs, test for cycle
